refactor : Refactor the loading.tsx file to incorporate email verification functionality.
