+++ src/app/auth/email/verify/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import VerifyEmail from "@/components/custom/email/verify/verify-email"
-import Loading from "@/components/ui/loading"
+import Loading from "@/app/auth/email/verify/loading"
 import { Suspense } from "react"
 
 export default function Verify() {
